+++ src/cpp/session/modules/SessionEnvironment.R
@@ -670,7 +670,7 @@
    .Call("rs_hasExternalPointer", object, nullPtr, PACKAGE = "(embedding)")
 })
 
-.rs.addFunction("describeObject", function(env, objName, isAltrep)
+.rs.addFunction("describeObject", function(env, objName)
 {
    obj <- get(objName, env)
 
@@ -772,7 +772,7 @@
          if (is.list(obj) ||  is.data.frame(obj) || isS4(obj) ||
              inherits(obj, c("data.table", "ore.frame", "cast_df", "xts", "DataFrame")))
          {
-            if (isAltrep)
+            if (.rs.isAltrep(obj))
             {
                # don't prefetch content for altreps
                val <- "NO_VALUE"