+++ NEWS.md
@@ -5,6 +5,7 @@
 #### RStudio
 
 - RStudio now highlights all keywords from the SQL 2023 standard in SQL documents (#15841)
+- RStudio now uses lobstr when computing object sizes (#15919)
 - Improved legibility of highlighted code when RStudio debugger is active
 - Added a user preference to disable showing the splash screen at startup (#15945)
 - The splash screen now closes when clicked with the mouse (#15614)